// src/api/models/CalendarConnectionCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CalendarConnectionCreate = {
    name: string;
    color: string;
    url: string;
    provider?: (string | null);
    enabled?: boolean;
    profile_id: number;
};