// src/api/models/CalendarConnectionList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CalendarConnectionRead } from './CalendarConnectionRead';
export type CalendarConnectionList = {
    calendar_connections?: Array<CalendarConnectionRead>;
    total: number;
    limit: number;
    offset: number;
};