// src/api/models/CalendarConnectionUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CalendarConnectionUpdate = {
    name?: (string | null);
    color?: (string | null);
    url?: (string | null);
    provider?: (string | null);
    enabled?: (boolean | null);
};