// src/api/models/CalendarEventList.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CalendarEventRead } from './CalendarEventRead';
export type CalendarEventList = {
    events?: Array<CalendarEventRead>;
    date: string;
    errors?: Array<string>;
};